Removed csgo  - it was a free game
Changed naming convention of icons

Cleaned up scss

Redid small bits of the schema
 - removed nickname field from users
 - removed user_games association between users and games

Initial work on authentication
 - added controller for users with new, show, edit, create and update actions
 - added controller for logins with new, create and destroy actions
 - added sign up form for registering (users#create)

+++ db/migrate/20211120143329_create_reviews.rb
@@ -3,7 +3,7 @@ class CreateReviews < ActiveRecord::Migration[6.1]
     create_table :reviews do |t|
       t.belongs_to :user
       t.belongs_to :game
-      t.decimal :rating
+      t.integer :rating
       t.text :content
 
       t.timestamps
